Removing Version() check in tests. (#130)
It is a reasonable expectation from our customers that `go test` should
execute without failure on master from their machines. At the moment,
there is a by design failure because the version matches the most
recently published tag.

Eventually, we should enforce this again, but not until we can do so
only for our own release pipeline. To ensure folks don't get snared on
this, we'll delete the offending test here.