Finished Text, Binary, and JSON for UnixTime (#135)
* Finished Text, Binary, and JSON for UnixTime

* Updating package ref to target Azure fork.

* Responding to review feedback

* Fixing example name typo

* Updating MarshalJSON to test on tolerance

Floats may differ slightly, while still returning exactly the same
nanosecond. If that scenario is ever encountered, we don't want the
tests too fail.

* Keep nanosecond resolution in JSON.

* Updating JSON Roundtrip for tolerance

Previously, JSONRoundTrip asserted that we are able to have granularity
to nanoseconds even when storing time as a float64 of nanoseconds. We
have run into some trouble with that, and must accept that if one needs
more precision, marshalling as Text or Binary will have to suffice over
encoding it as a JSON number.

* Cleaning up constructors to not clutter API space