Stop retriable request from adding a non-nil Body (#726)
When azure is responding with an error the retriable request that is generated ends up getting a non nil noOpcloser added to it with a reader interface. Ultimately this ends up sending a GET request to Azure that has a post body. As the http spec says that the body SHOULD be ignored this is not a critical error, but for people using custom transports this is problematic. It's also sending unessecary bytes.